# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de SalonFlow",
        "description": "API REST para gestión de usuarios, servicios y citas",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/AppSalon/public/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/tokens": {
            "get": {
                "tags": [
                    "Gestión de Tokens"
                ],
                "summary": "Listar tokens activos",
                "description": "Retorna todos los tokens activos del usuario autenticado",
                "operationId": "listTokens",
                "responses": {
                    "200": {
                        "description": "Lista de tokens obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tokens obtenidos exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "auth_token"
                                                    },
                                                    "last_used_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/tokens/{id}": {
            "delete": {
                "tags": [
                    "Gestión de Tokens"
                ],
                "summary": "Revocar token específico",
                "description": "Revoca un token específico del usuario autenticado",
                "operationId": "revokeToken",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del token a revocar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token revocado exitosamente"
                    },
                    "404": {
                        "description": "Token no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}